okay NOW speed is actually a float
had to change the interface and stuff

+++ Assets/BrackeysGameJam/Scripts/Pickups.cs
@@ -21,7 +21,7 @@ namespace Boxfriend
         #endregion
 
         #region IInteractable
-        public int SpeedChange { get { return _speedChange; } }
+        public float SpeedChange { get { return _speedChange; } }
         public int HealthChange { get { return _healthChange; } }
         #endregion
